+++ src/Events/IceHawkWasInitializedEvent.php
@@ -17,17 +17,11 @@ final class IceHawkWasInitializedEvent implements CarriesEventData
 	/** @var ProvidesRequestInfo */
 	private $requestInfo;
 
-	/**
-	 * @param ProvidesRequestInfo $requestInfo
-	 */
 	public function __construct( ProvidesRequestInfo $requestInfo )
 	{
 		$this->requestInfo = $requestInfo;
 	}
 
-	/**
-	 * @return ProvidesRequestInfo
-	 */
 	public function getRequestInfo() : ProvidesRequestInfo
 	{
 		return $this->requestInfo;