+++ src/Events/InitializingIceHawkEvent.php
@@ -13,17 +13,11 @@ final class InitializingIceHawkEvent implements CarriesEventData
 	/** @var ProvidesRequestInfo */
 	private $requestInfo;
 
-	/**
-	 * @param ProvidesRequestInfo $requestInfo
-	 */
 	public function __construct( ProvidesRequestInfo $requestInfo )
 	{
 		$this->requestInfo = $requestInfo;
 	}
 
-	/**
-	 * @return ProvidesRequestInfo
-	 */
 	public function getRequestInfo() : ProvidesRequestInfo
 	{
 		return $this->requestInfo;