+++ src/Events/ReadRequestWasHandledEvent.php
@@ -19,9 +19,6 @@ final class ReadRequestWasHandledEvent implements CarriesEventData
 	/** @var ProvidesReadRequestData */
 	private $request;
 
-	/**
-	 * @param ProvidesReadRequestData $request
-	 */
 	public function __construct( ProvidesReadRequestData $request )
 	{
 		$this->request = $request;
@@ -29,11 +26,11 @@ final class ReadRequestWasHandledEvent implements CarriesEventData
 
 	public function getRequestInfo() : ProvidesRequestInfo
 	{
-		return $this->request->getRequestInfo();
+		return $this->request->getInfo();
 	}
 
-	public function getInputData() : ProvidesReadRequestInputData
+	public function getRequestInput() : ProvidesReadRequestInputData
 	{
-		return $this->request->getInputData();
+		return $this->request->getInput();
 	}
 }