+++ tests/Unit/Events/HandlingReadRequestEventTest.php
@@ -21,6 +21,6 @@ class HandlingReadRequestEventTest extends \PHPUnit_Framework_TestCase
 		$event = new HandlingReadRequestEvent( $readRequest );
 
 		$this->assertSame( $requestInfo, $event->getRequestInfo() );
-		$this->assertSame( $requestInput, $event->getInputData() );
+		$this->assertSame( $requestInput, $event->getRequestInput() );
 	}
 }