+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com).
 
+## [Unreleased]
+
+### Changed
+
+- Renamed accessors for RequestInfo and InputData on `$request` object to `$request->getInfo()` and `$request->getInput()` 
+
 ## [2.0.0-rc4] - 2016-07-13
 
 **CAUTION** This release has major interface changes and is not backward compatible with prior versions.
@@ -142,6 +148,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a CH
 
 - First release
 
+[Unreleased]: https://github.com/icehawk/icehawk/compare/v2.0.0-rc4...development
 [2.0.0-rc4]: https://github.com/icehawk/icehawk/compare/v1.4.2...v2.0.0-rc4
 [1.4.2]: https://github.com/icehawk/icehawk/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/icehawk/icehawk/compare/v1.4.0...v1.4.1